% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMER.data.R
\docType{data}
\name{EPIC.hg38.manifest}
\alias{EPIC.hg38.manifest}
\title{A GRanges containing hg38 annotation with suggested overall masking for EPIC platform}
\format{A GRanges with 866895 elements}
\usage{
EPIC.hg38.manifest
}
\description{
A GRanges containing hg38 annotation with suggested overall masking for EPIC platform
}
\examples{
\dontrun{
data("EPIC.hg38.manifest")
}
}
\keyword{internal}
