% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMER.data.R
\docType{data}
\name{EPIC.manifest.hg38}
\alias{EPIC.manifest.hg38}
\title{A GRanges containing hg38 annotation with suggested overall masking for EPIC platform}
\format{A GRanges with 866895 elements}
\usage{
EPIC.manifest.hg38
}
\description{
A GRanges containing hg38 annotation with suggested overall masking for EPIC platform
}
\examples{
\dontrun{
data("EPIC.manifest.hg38")
}
}
\keyword{internal}
