% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMER.data.R
\docType{data}
\name{EPIC.manifest}
\alias{EPIC.manifest}
\title{A GRanges containing hg19 annotation with suggested overall masking for EPIC platform}
\format{A GRanges with 866895 elements}
\usage{
EPIC.manifest
}
\description{
A GRanges containing hg19 annotation with suggested overall masking for EPIC platform
}
\examples{
data("EPIC.manifest")
}
\keyword{internal}
