% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMER.data.R
\docType{data}
\name{hm450.hg38.manifest}
\alias{hm450.hg38.manifest}
\title{A GRanges containing hg38 annotation with suggested overall masking for hm450 platform}
\format{A GRanges with 485577 elements}
\usage{
hm450.hg38.manifest
}
\description{
A GRanges containing hg38 annotation with suggested overall masking for hm450 platform
}
\examples{
\dontrun{
data("hm450.hg38.manifest")
}
}
\keyword{internal}
