% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMER.data.R
\docType{data}
\name{hm450.manifest.hg38}
\alias{hm450.manifest.hg38}
\title{A GRanges containing hg38 annotation with suggested overall masking for hm450 platform}
\format{A GRanges with 485577 elements}
\description{
A GRanges containing hg38 annotation with suggested overall masking for hm450 platform
}
\examples{
data("hm450.manifest.hg38")
}
\keyword{internal}
