% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMER.data.R
\docType{data}
\name{hocomoco.table}
\alias{hocomoco.table}
\title{Table parsed from hocomoco v11}
\format{A dataframe with 771 rows (motifs) and 20 columns}
\usage{
hocomoco.table
}
\description{
Table parsed from hocomoco v11
}
\keyword{internal}
