% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMER.data.R
\docType{data}
\name{Human_genes__GRCh37_p13}
\alias{Human_genes__GRCh37_p13}
\title{A matrix containing ENSEMBL hg19 gene metadata accessed using biomart
This data is used if ensembl cannot be reached}
\format{A matrix with metadata for 60482 genes}
\usage{
Human_genes__GRCh37_p13
}
\description{
A matrix containing ENSEMBL hg19 gene metadata accessed using biomart
This data is used if ensembl cannot be reached
}
\keyword{internal}
