% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMER.data.R
\docType{data}
\name{Human_genes__GRCh38_p12}
\alias{Human_genes__GRCh38_p12}
\title{A matrix containing ENSEMBL hg38 gene metadata accessed using biomart
This data is used if ensembl cannot be reached}
\format{A matrix with metadata for 58639 genes}
\usage{
Human_genes__GRCh38_p12
}
\description{
A matrix containing ENSEMBL hg38 gene metadata accessed using biomart
This data is used if ensembl cannot be reached
}
\keyword{internal}
