% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMER.data.R
\docType{data}
\name{Human_genes__GRCh38_p12__tss}
\alias{Human_genes__GRCh38_p12__tss}
\title{A matrix containing ENSEMBL hg38 transcripts metadata accessed using biomart
This data is used if ensembl cannot be reached}
\format{A matrix with metadata for 208423 transcripts}
\usage{
Human_genes__GRCh38_p12__tss
}
\description{
A matrix containing ENSEMBL hg38 transcripts metadata accessed using biomart
This data is used if ensembl cannot be reached
}
\keyword{internal}
