% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMER.data.R
\docType{data}
\name{mae.promoter}
\alias{mae.promoter}
\title{A MultiAssayExperiment containing
DNA methylation data: 16 promoter probes from platform 450K 
Gene Expression data: 3808 genes
for 234 samples from TCGA-LUSC.
This data is used in the examples of ELMER package}
\format{A MultiAssayExperiment for 234 Samples (8 normal samples, 226 Primary solid tumor)}
\usage{
mae.promoter
}
\description{
A MultiAssayExperiment containing
DNA methylation data: 16 promoter probes from platform 450K 
Gene Expression data: 3808 genes
for 234 samples from TCGA-LUSC.
This data is used in the examples of ELMER package
}
\examples{
\dontrun{
data("elmer.data.example.promoter")
}
}
\keyword{internal}
