% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMER.data.R
\docType{data}
\name{motif.relevant.TFs}
\alias{motif.relevant.TFs}
\title{A list with TF that might bind to each motif}
\format{A List with 641 elements}
\description{
A list with TF that might bind to each motif
}
\examples{
data("motif.relevant.TFs")
}
\keyword{internal}
