% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMER.data.R
\docType{data}
\name{Probes.motif.hg38.EPIC}
\alias{Probes.motif.hg38.EPIC}
\title{A matrix with 1 if the probe (row) has a motif (column)}
\format{A matrix with  838881 rows and 640 columns}
\description{
A matrix with 1 if the probe (row) has a motif (column)
}
\examples{
data("Probes.motif.hg38.EPIC")
}
\keyword{internal}
