\name{Probes.motif}
\alias{Probes.motif}
\docType{data}
\title{
  Motif occurences within -/+100bp of probe sites on HM450K array. 
}
\description{
  FIMO was used with a p-value < 0.0001 to scan a +/- 100bp region around each probe 
  using Factorbook motif position weight matrices (PWMs) and Jasper core human 
  motif PWMs generated from the R package MotifDb.This data set is used in 
  get.enriched.motif function in \pkg{ELMER}
}
\usage{data("Probes.motif")}
\format{
  A matrix with 485512 row and 91 column. Each row is each probe regions and
  each column is motif from Factorbook and Jasper. The value 1 indicates the occurrence 
  of a motif in a particular probe and 0 means no occurrence.
}
\value{
  A matrix with 485512 row and 91 column. Each row is each probe regions and
  each column is motif from Factorbook and Jasper. The value 1 indicates the occurrence 
  of a motif in a particular probe and 0 means no occurrence.
}
\references{
1. Wang J, Zhuang J, Iyer S, Lin XY, Greven MC, Kim BH, Moore J, Pierce BG, 
Dong X, Virgil D, et al: Factorbook.org: a Wiki-based database for transcription 
factor-binding data generated by the ENCODE consortium. Nucleic Acids Res 2013, 
41:D171-176.
2. Grant CE, Bailey TL, Noble WS: FIMO: scanning for occurrences of a given motif. 
Bioinformatics 2011, 27:1017-1018.
3. Wang J, Zhuang J, Iyer S, Lin X, Whitfield TW, Greven MC, Pierce BG, Dong X, 
Kundaje A, Cheng Y, et al: Sequence features and chromatin structure around the 
genomic regions bound by 119 human transcription factors. Genome Res 2012,
22:1798-1812.
4. Shannon P: MotifDb: An annotated collection of Protein-DNA binding sequence 
motifs. Bioconductor 2014, R package version 1.8.0.
5. Yao L, Shen H, Laird PW, Farnham PJ,Berman BP: Inferring Regulatory Element 
Landscapes and Transcription Factor Networks from Cancer Methylomes. in revision of
Genome Biology
}
\examples{
data("Probes.motif")
dim(Probes.motif)
str(Probes.motif)
}
\keyword{datasets}
