% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMER.data.R
\docType{data}
\name{TF.family}
\alias{TF.family}
\title{A list of 641 motifs with TF families (with similar bidings) from TFClass
Created with the following function from ELMER pacakge 
TF.family <-  createMotifRelevantTfs()}
\format{A list of 641 motifs with TF families (with similar bidings)}
\usage{
TF.family
}
\description{
A list of 641 motifs with TF families (with similar bidings) from TFClass
Created with the following function from ELMER pacakge 
TF.family <-  createMotifRelevantTfs()
}
\keyword{internal}
