% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMER.data.R
\docType{data}
\name{TF.subfamily}
\alias{TF.subfamily}
\title{A list of 641 motifs with TF subfamilies (with similar bidings) from TFClass
Created with the following function from ELMER pacakge 
TF.family <-  createMotifRelevantTfs("subfamily")}
\format{A list of 641 motifs with TF subfamilies (with similar bidings)}
\description{
A list of 641 motifs with TF subfamilies (with similar bidings) from TFClass
Created with the following function from ELMER pacakge 
TF.family <-  createMotifRelevantTfs("subfamily")
}
\keyword{internal}
