% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMER.data.R
\docType{data}
\name{Union.enhancer.hg38}
\alias{Union.enhancer.hg38}
\title{A GRanges with enhancer regions}
\format{A GRRanges with 571084 elements}
\description{
A GRanges with enhancer regions
}
\examples{
data("Union.enhancer.hg38")
}
\keyword{internal}
