% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMER.data.R
\docType{package}
\name{ELMER.data}
\alias{ELMER.data}
\alias{ELMER.data-package}
\title{Data for ELMER package}
\description{
ELMER is package using DNA methylation to 
identify enhancers, and correlates enhancer state with expression of nearby genes 
to identify one or more transcriptional targets. Transcription factor (TF) binding 
site analysis of enhancers is coupled with expression analysis of all TFs to 
infer upstream regulators. ELMER.data provide 2 necessary data for 
ELMER analysis:
\itemize{
  \item Probes.motif: motif occurences within -/+250bp of probe sites on HM450K/EPIC array aligned against hg19/hg38.
  \item Union.enhancer: A comprehensive list of genomic strong enhancers.
  \item DNA methylation platform manifest: from http://zwdzwd.github.io/InfiniumAnnotation
}
    For more information how to create these objects please read the vignette of this package with the
    follwing command: \code{browseVignettes("ELMER.data")}
}
\examples{
data("Union.enhancer.hg38")
data("Union.enhancer.hg19")
data("motif.relavent.TFs")
data("Probes.motif.hg38.EPIC")
data("Probes.motif.hg19.EPIC")
data("Probes.motif.hg38.450K")
data("Probes.motif.hg19.450K")
data("EPIC.manifest")
data("EPIC.manifest.hg38")
data("hm450.manifest")
data("hm450.manifest.hg38")
}
\keyword{utilities}
