% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMER.data.R
\docType{data}
\name{human.TF}
\alias{human.TF}
\title{Table parsed from Lambert, Samuel A., et al. "The human transcription factors." Cell 172.4 (2018): 650-665.}
\format{A dataframe with 1639 rows (motifs) and 27 columns}
\usage{
human.TF
}
\description{
Table parsed from Lambert, Samuel A., et al. "The human transcription factors." Cell 172.4 (2018): 650-665.
}
\keyword{internal}
